sports/ffbaseball/requirements.txt
espn_api>=0.12.0
click>=8.1.7
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0 